Convert _currentIndex to double for DotsIndicator

+++ lib/screens/home_screen.dart
@@ -226,7 +226,7 @@ class ImageSliderIndicator extends StatelessWidget {
       padding: const EdgeInsets.only(top: 8.0),
       child: DotsIndicator(
         dotsCount: carouselImages.length,
-        position: _currentIndex,
+        position: _currentIndex.toDouble(),
         decorator: DotsDecorator(
           // ACTIVE DOT (solid dot inside a hole)
           activeColor: const Color.fromRGBO(141, 199, 63, 1),
